Major release: Bug fixes, user improvements and code optimisation.
New:
A new look for the `--help' manual page.
An option to report 'short StORFs' that are under 100nts.
A new filtering option is available to filter out StORFs that do not include at least one of the three canonical start codons ('ATG', 'GTG', 'TTG') - Most useful for short StORF reporting.

Code Optimisation and fixes:
General code optimisations (including the removal of unnecessary sorting of StORFs during the filtering stage).
Bug fixes that include the incorrect filtering of Con-StORFs due to a sequence length bug.
Fixes to the StORF-Reporter PROKKA output to now report the same StORF ID between the GFF and FASTA and to only output the DNA sequences of the PROKKA and StORF genes.

Misc:
StORF-Reporter version now in the header of all GFF files produced.